User services documentation.

+++ src/services/User.service.js
@@ -1,54 +1,94 @@
 const userModel = require('./../models/User.model');
-const fr = require('./../utils/FormatResponse');
 
 async function login(dto) {
     const user = await userModel.findOne({ where: { email: dto.email } });
 
-    if (user === null) return fr.responseError("E-mail não encontrado.");
+    if (user === null) throw new Error('E-mail não encontrado.');
 
-    if (user.password !== dto.password) return fr.responseError("Senha incorreta.");
+    if (user.password !== dto.password) throw new Error('Senha incorreta.');
 
-    return fr.responseSucces(user);
+    return {
+        ...user.dataValues,
+        password: null
+    };
 }
 
 async function createUser(dto) {
     const findEmail = await userModel.count({ where: { email: dto.email } });
 
-    if (findEmail >= 1) return fr.responseError("E-mail já cadastrado no sistema.");
+    if (findEmail >= 1) throw new Error('E-mail já cadastrado no sistema.');
 
     const user = await userModel.create(dto);
 
-    return fr.responseSucces(user);
+    return {
+        ...user.dataValues,
+        password: null
+    };
 }
 
 async function getUserById(id) {
     const user = await userModel.findOne({
-        attributes: {
-            exclude: ['password']
-        },
         where: {
             idUser: parseInt(id)
         }
     });
-    return user;
+
+    if (user === null) throw new Error('Usuário não encontrado.');
+
+    return {
+        ...user.dataValues,
+        password: null
+    };
 }
 
 async function updateUser(id, dto) {
-    const user = await userModel.update(dto, {
+    const userData = await userModel.findOne({
         where: {
             idUser: parseInt(id)
         }
     });
-    return user;
+
+    if (userData === null) throw new Error('Usuário não encontrado.');
+
+    if (userData.password != dto.password) throw new Error('Senha incorreta, autorização negada.');
+
+    const uptObj = {
+        name: dto.name != userData.name ? dto.name : userData.name,
+        photo: dto.photo != userData.photo ? dto.photo : userData.photo,
+        nickname: dto.nickname != userData.nickname ? dto.nickname : userData.nickname,
+        dateOfBirth: dto.dateOfBirth != userData.dateOfBirth ? dto.dateOfBirth : userData.dateOfBirth,
+        email: dto.email != userData.email ? dto.email : userData.email,
+        nationality: dto.nationality != userData.nationality ? dto.nationality : userData.nationality,
+        phone: dto.phone != userData.phone ? dto.phone : userData.phone,
+    }
+
+    const user = await userModel.update(uptObj, {
+        where: {
+            idUser: parseInt(id)
+        }
+    });
+
+    if (user === null) throw new Error('Não foi possível atualizar usuário.');
+
+    uptObj.password = null;
+
+    return uptObj;
 }
 
 async function deleteUser(id) {
+    const findUser = await userModel.count({ where: { idUser: id } });
+
+    if (findUser === 0) throw new Error('Usuário não encontrado.');
+
     const user = await userModel.destroy({
         where: {
             idUser: parseInt(id)
         }
     });
-    return user;
+
+    if (user === null) throw new Error('Não foi possível deletar usuário.');
+
+    return 'Usuário deletado com sucesso.';
 }
 
 module.exports = {
